feat(address-service): ♻️ DOMA-11602 The address resolution utilities and tests were moved to separated space
